docs: refactor custom-caching example
